Support reverse projection during geometry editing for layers that "reproject to declared" srs
Added support for projecting to the "native" CRS when modifying geometry in a layer that is setup using "Reproject to declared" projection policy.  It was found that editing the geometry (through wfs in this case) would result in coordinates in the "declared" srs landing in the native feature store.
(Tony)